Refactor auth policy forms, add useDebounce hook
Extracted JWT and OIDC config sections into separate components for better modularity in AuthenticationPolicyEdit. Added a reusable useDebounce hook and applied it to Certificates, Clusters, and Routes search fields. Introduced a shared form data type for authentication policies. Improved code organization and search performance across list pages.
